refactor(events): move event utils into events namespace
