+++ src/events/utils/isSelfEvent.test.ts
