+++ src/events/utils/isSelfEvent.ts
